rename scripts
