+++ .github/actions/test_ya/action.yml
@@ -565,7 +565,9 @@ runs:
                 MESSAGE="${MESSAGE}${SEPARATOR}Unmuted $LINE_COUNT_unmute [tests](${PUBLIC_DIR_URL}/mute_info/3_unmuted_tests.txt)"
               fi
             fi
-            printf "$MESSAGE" | .github/scripts/tests/comment-pr.py --color orange
+            if [ -n "$MESSAGE" ]; then
+              printf "$MESSAGE" | .github/scripts/tests/comment-pr.py --color orange
+            fi
           fi
